+++ Momentarily/Momentarily.Web/Areas/Frontend/Controller/HomeController.cs
@@ -121,42 +121,42 @@ namespace Momentarily.Web.Areas.Frontend.Controller
             return View(fAQVM);
         }
 
-        [HttpGet]
-        public ActionResult ContactUs()
-        {
-            ViewBag.Title = "Help Centre - Contact Us | momentarily";
-            ViewBag.Description = "Contact us for issues regarding using the momentarily platform.";
-            var model = new ContactUsEntry();
+        //[HttpGet]
+        //public ActionResult ContactUs()
+        //{
+        //    ViewBag.Title = "Help Centre - Contact Us | momentarily";
+        //    ViewBag.Description = "Contact us for issues regarding using the momentarily platform.";
+        //    var model = new ContactUsEntry();
     
-            var shape = _shapeFactory.BuildShape(null, model, PageName.Home.ToString());
-            return DisplayShape(shape);
-        }
-        [HttpPost]
-        public ActionResult ContactUs(Shape<ContactUsEntry> shape)
-        {
-            if (!this.IsCaptchaValid("Validate your captcha"))
-            {
-                ViewBag.ErrorMessage = "Invalid Captcha";
-                return DisplayShape(shape);
-            }
+        //    var shape = _shapeFactory.BuildShape(null, model, PageName.Home.ToString());
+        //    return DisplayShape(shape);
+        //}
+        //[HttpPost]
+        //public ActionResult ContactUs(Shape<ContactUsEntry> shape)
+        //{
+        //    if (!this.IsCaptchaValid("Validate your captcha"))
+        //    {
+        //        ViewBag.ErrorMessage = "Invalid Captcha";
+        //        return DisplayShape(shape);
+        //    }
 
-            if (ModelState.IsValid)
-            {
-                var result = _sendMessageService.SendEmailContactUs(shape.ViewModel);
+        //    if (ModelState.IsValid)
+        //    {
+        //        var result = _sendMessageService.SendEmailContactUs(shape.ViewModel);
 
-                shape.IsError = !result;
-                if (result)
-                {
-                    shape.ViewModel = new ContactUsEntry();
-                    ViewBag.IsEmailSent = "True";
-                }
-                else
-                {
-                    ViewBag.IsEmailSent = "False";
-                }
-            }
-            return DisplayShape(shape);
-        }
+        //        shape.IsError = !result;
+        //        if (result)
+        //        {
+        //            shape.ViewModel = new ContactUsEntry();
+        //            ViewBag.IsEmailSent = "True";
+        //        }
+        //        else
+        //        {
+        //            ViewBag.IsEmailSent = "False";
+        //        }
+        //    }
+        //    return DisplayShape(shape);
+        //}
         public ActionResult HowItWorks()
         {
             ViewBag.Title = "How momentarily Works | Share & Earn; Borrow & Save | momentarily";
