+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/Account/Register.cshtml
@@ -4,7 +4,7 @@
 }
 @*<script src="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.5/js/select2.js"></script>*@
 @*<script src="https://stackpath.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js" integrity="sha384-aJ21OjlMXNL5UyIl/XNwTMqvzeRMZH2w8c5cRVpzpU8Y5bApTppSuUkhZXN0VxHd" crossorigin="anonymous"></script>*@
-<script src="https://www.google.com/recaptcha/api.js"></script>
+<script src="https://www.google.com/recaptcha/api.js" async defer></script>
 
 <div class="content mainClass" data-ng-controller="RegisterController as account">
     <div class="container-full">
