+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/Home/ContactUs.cshtml
@@ -106,7 +106,7 @@
                 <div class="subscribeInner">
                     <h4>Subscribe to Our Newsletter</h4>
                     <p>For latest updates and news</p>
-                    <!---->
+                    
                     <div class="input-group">
                         <input type="email" id="subscribeEmail" class="form-control" placeholder="Enter your email">
                         <span class="input-group-btn">
@@ -115,7 +115,7 @@
                     </div>
                     <span class="subscribe_msg red" id="subscribe_msg_error" style="display:none; text-align:left;"><i class="fa fa-exclamation-triangle" aria-hidden="true"></i> Invalid email address.</span>
                     <span class="subscribe_msg green" id="subscribe_msg_error2" style="display:none; text-align:left;"><i class="fa fa-exclamation-triangle" aria-hidden="true"></i> Please enter email address.</span>
-                    <!---->
+                    
                 </div>
             </div>
         </div>
@@ -133,7 +133,7 @@
     </div>
 </section>
 
-<!--footer section-->
+footer section
 <footer class="footer">
     <div class="container-fluid">
         <div class="row">
@@ -152,9 +152,9 @@
                                 <li><a href="@Url.Action("Index", "Home")" id="btn_home" "><i class="fa fa-caret-right" aria-hidden="true"></i> Home</a></li>
                                 <li><a href="@Url.Action("AboutUs", "Home")"><i class="fa fa-caret-right" aria-hidden="true"></i> About Us</a></li>
                                 <li><a href="@Url.Action("FAQ", "Home")"><i class="fa fa-caret-right" aria-hidden="true"></i>  FAQ</a></li>
-                                <!--  <li><a href="#"><i class="fa fa-caret-right" aria-hidden="true"></i> Videos</a></li>
+                                  <li><a href="#"><i class="fa fa-caret-right" aria-hidden="true"></i> Videos</a></li>
                                    <li><a href="#"><i class="fa fa-caret-right" aria-hidden="true"></i> Blog</a></li>
-                                   <li><a href="/contactus"><i class="fa fa-caret-right" aria-hidden="true"></i> Help</a></li> -->
+                                   <li><a href="/contactus"><i class="fa fa-caret-right" aria-hidden="true"></i> Help</a></li> 
                             </ul>
                         </div>
                     </div>
@@ -165,7 +165,7 @@
                                 <li><a href="@Url.Action("HowItWorks", "Home")"><i class="fa fa-caret-right" aria-hidden="true"></i> How it Works </a></li>
                                 <li><a href="@Url.Action("Create", "Listing")"><i class="fa fa-caret-right" aria-hidden="true"></i> Share </a></li>
                                 <li><a href="" id="becomeBorrower"><i class="fa fa-caret-right" aria-hidden="true"></i> Borrow </a></li>
-                                <li><a href="@Url.Action("ContactUs", "Home")"><i class="fa fa-caret-right" aria-hidden="true"></i>  Help</a></li>
+                                <li><a href="@Url.Action("sendhelp", "Home")"><i class="fa fa-caret-right" aria-hidden="true"></i>  Help</a></li>
                             </ul>
                         </div>
                     </div>
@@ -195,7 +195,7 @@
         </div>
     </div>
 </footer>
-<!---->
+
 <div class="copyright">
     <div class="container">
         <div class="row">
