+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/Home/FAQ.cshtml
@@ -61,7 +61,7 @@
                 .replace('$ set up your momentarily member account $', '<a target="_blank" href="@Url.Action("Register", "Account")">setup your momentarily member account </a>')
                 .replace('$ Become a Sharer $', '<a target="_blank" href="@Url.Action("Create", "Listing")">Become a Sharer </a>')
                 .replace('$ momentarily bold $', '<b>momentarily </b>')
-                .replace('$ contact $', '<a target="_blank" href="@Url.Action("ContactUs", "Home")">contact</a>')
+                .replace('$ contact $', '<a target="_blank" href="@Url.Action("index", "sendhelp")">contact</a>')
                 .replace('$contactus$', '<a target="_blank" href="@Url.Action("ContactUs", "Home")">contact us</a>')
                 .replace('$Owner$', '<a target="_blank" href="@Url.Action("Index", "Listing")">Owner</a>')
                 .replace('$Borrower$', '<a target="' + target + '" href="" class="becomeBorrower" >Borrower</a>')
