+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/Shared/_Footer.cshtml
@@ -30,7 +30,8 @@
                                 <li><a href="@Url.Action("HowItWorks", "Home")"><i class="fa fa-caret-right" aria-hidden="true"></i> How it Works </a></li>
                                 <li><a href="@Url.Action("Create", "Listing")"><i class="fa fa-caret-right" aria-hidden="true"></i> Share </a></li>
                                 <li><a href="" id="becomeBorrower"><i class="fa fa-caret-right" aria-hidden="true"></i> Borrow </a></li>
-                                <li><a href="@Url.Action("ContactUs", "Home")"><i class="fa fa-caret-right" aria-hidden="true"></i>  Help</a></li>
+                                @*<li><a href="@Url.Action("ContactUs", "Home")"><i class="fa fa-caret-right" aria-hidden="true"></i>  Help</a></li>*@
+                                <li><a href="@Url.Action("index", "sendhelp")"><i class="fa fa-caret-right" aria-hidden="true"></i>  Help</a></li>
                             </ul>
                         </div>
                     </div>
