+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/Shared/_LayoutNew.cshtml
@@ -100,15 +100,17 @@
                                         <li><a id="search_button" href="#"><img src="~/Content/TemplateImg/OtherImages/searchicon.png" alt="image" /> </a></li>
                                         <li class="active"><a href="@Url.Action("HowItWorks", "Home")">How it Works</a></li>
                                         <li><a href="@Url.Action("FAQ", "Home")">Faq</a></li>
-                                        <li><a href="@Url.Action("ContactUs", "Home")">Help</a></li>
-                                        <li class="dropdown dropNav covidMenu afterLoginCovidMenu">
+                                        @*<li><a href="@Url.Action("ContactUs", "Home")">Help</a></li>*@
+                                        <li><a href="@Url.Action("index", "sendhelp")">Help</a></li>
+                                        
+                                        @*<li class="dropdown dropNav covidMenu afterLoginCovidMenu">
                                             <a class="dropdown-toggle" data-toggle="dropdown" href="#" aria-expanded="false">
                                                 <span> COVID-19</span> <span class="fa fa-angle-down"></span>
                                             </a>
                                             <ul class="dropdown-menu">
                                                 <li class="@((IsActiveLink("Index", "covid")||IsActiveLink("PlaceOrder", "covid")) ? "active" : "")"><a href="@Url.Action("Index", "covid")">COVID-19 Products</a></li>
                                             </ul>
-                                        </li>
+                                        </li>*@
 
 
                                         <li class="dropdown dropNav">
@@ -164,15 +166,15 @@
                                     <ul class="nav navbar-nav navbar-right">
                                         <li class="active"><a href="@Url.Action("HowItWorks", "Home")">How it Works</a></li>
                                         <li><a href="@Url.Action("FAQ", "Home")">Faq</a></li>
-                                        <li><a href="@Url.Action("ContactUs", "Home")">Help</a></li>
-                                        <li class="dropdown dropNav covidMenu">
-                                            <a class="dropdown-toggle" data-toggle="dropdown" href="#" aria-expanded="false">
-                                                <span> COVID-19</span> <span class="fa fa-angle-down"></span>
-                                            </a>
-                                            <ul class="dropdown-menu">
-                                                <li class="@((IsActiveLink("Index", "covid")||IsActiveLink("PlaceOrder", "covid")) ? "active" : "")"><a href="@Url.Action("Index", "covid")">COVID-19 Products</a></li>
-                                            </ul>
-                                        </li>
+                                        <li><a href="@Url.Action("index", "sendhelp")">Help</a></li>
+                                        @*<li class="dropdown dropNav covidMenu">
+            <a class="dropdown-toggle" data-toggle="dropdown" href="#" aria-expanded="false">
+                <span> COVID-19</span> <span class="fa fa-angle-down"></span>
+            </a>
+            <ul class="dropdown-menu">
+                <li class="@((IsActiveLink("Index", "covid")||IsActiveLink("PlaceOrder", "covid")) ? "active" : "")"><a href="@Url.Action("Index", "covid")">COVID-19 Products</a></li>
+            </ul>
+        </li>*@
 
 
                                         <li><a href="@QuickUrl.RegisterUrl()"><span class="fa fa-user"></span> Sign up</a></li>
@@ -180,7 +182,7 @@
 
                                         <li class="becomeNav"><a href="@Url.Action("Create", "Listing")" class="btn btn-primary btn-sm"> Share </a></li>
                                         @*<li class="becomeNav becomeBorrowerCls"><a href="/Search/Map" class="btn btn-primary btn-sm"> Borrow </a>
-                                            </li>*@
+        </li>*@
                                         <li class="becomeNav becomeBorrowerCls" id="becomeBorrower"><a class="btn btn-primary btn-sm"> Borrow </a></li>
                                     </ul>
                                 </div>
@@ -225,8 +227,9 @@
             <ul>
                 <li class="active"><a href="@Url.Action("HowItWorks", "Home")">How it Works</a></li>
                 <li><a href="@Url.Action("FAQ", "Home")">Faq</a></li>
-                <li><a href="@Url.Action("ContactUs", "Home")">Help</a></li>
-                <li class="dropdown dropNav covidMenu">
+                @*<li><a href="@Url.Action("ContactUs", "Home")">Help</a></li>*@
+                <li><a href="@Url.Action("index", "sendhelp")">Help</a></li>
+                @*<li class="dropdown dropNav covidMenu">
                     <a href="#">
                         <span> COVID-19</span> <span role="button" class="offcanvas-menu-toggler collapsed covidDDMenu" data-toggle="collapse" data-target="#covid-menu-02265" aria-expanded="false" aria-controls="collapse-menu-02">
                             <i class="fa fa-plus"></i><i class="fa fa-minus"></i>
@@ -235,7 +238,7 @@
                     <ul class="dropdown-menu" id="covid-menu-02265" aria-expanded="false">
                         <li class="@((IsActiveLink("Index", "covid")||IsActiveLink("PlaceOrder", "covid")) ? "active" : "")"><a href="@Url.Action("Index", "covid")">COVID-19 Products</a></li>
                     </ul>
-                </li>
+                </li>*@
                 @if (!User.Identity.IsAuthenticated && !(User.IsAdmin))
                 {
                     <li><a href="@QuickUrl.RegisterUrl()"><span class="fa fa-user"></span> Sign up</a></li>
