+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/User/UserEmailConfirm.cshtml
@@ -16,7 +16,7 @@
                         {
                             <div class="alerts">
                                 <div class="alert-error">
-                                    Your Email Id was not verified.
+                                    Your email was not verified.
                                 </div>
                             </div>
                         }
@@ -37,7 +37,7 @@
                             </div>
                         }
                         
-                       Please, click the button below to activate your account.
+                       Please click the button below to activate your account.
                         @using (@Html.BeginForm("UserEmailConfirmResend", "User", FormMethod.Post))
                         {
                             <button type="submit" id="send" class="btn btn-green send-button">Resend Verification</button>
