+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/covid/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<Momentarily.ViewModels.Models.CovidGoodViewModel>
+﻿<!--@model IEnumerable<Momentarily.ViewModels.Models.CovidGoodViewModel>
 @{
 
     Layout = "~/Areas/Frontend/Views/Shared/_LayoutNew.cshtml";
@@ -90,16 +90,16 @@
 
 <div class="modal fade covidPopup" tabindex="-1" role="dialog" id="covidpopup" aria-labelledby="myLargeModalLabel" aria-hidden="true">
     <div class="modal-dialog">
-        <div class="modal-content">
+        <div class="modal-content">-->
 
             <!-- Modal Header -->
-            <div class="modal-header">
+            <!--<div class="modal-header">
                 <h3>Covid-19</h3>
                 <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-            </div>
+            </div>-->
 
             <!-- Modal Body -->
-            <div class="modal-body">
+            <!--<div class="modal-body">
                 <div class="body-message">
                     <h5>We support our heroes during these rough times. Please read how your purchases will help go towards those essential workers and volunteers in our local communities.</h5>
                 </div>
@@ -535,4 +535,4 @@
         window.location.href = '/covid/PlaceOrder/' + id;
     }
 
-</script>
+</script>-->
