+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/covid/OrderCancelled.cshtml
@@ -1,4 +1,4 @@
-﻿@model Momentarily.ViewModels.Models.OrderPayResult
+﻿@*@model Momentarily.ViewModels.Models.OrderPayResult
 @{
     ViewBag.Title = "Pay";
     Layout = "~/Areas/Frontend/Views/Shared/_LayoutNew.cshtml";
@@ -72,4 +72,4 @@
         background-color: #fff;
         border-color: #cf2d1e;
     }
-</style>
+</style>*@
