+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/covid/Pay.cshtml
@@ -1,4 +1,4 @@
-﻿@model Momentarily.ViewModels.Models.OrderPayResult
+﻿@*@model Momentarily.ViewModels.Models.OrderPayResult
 @{
     ViewBag.Title = "Pay";
     Layout = "~/Areas/Frontend/Views/Shared/_LayoutNew.cshtml";
@@ -76,4 +76,4 @@
         background-color: #fff;
         border-color: #cf2d1e;
     }
-</style>
+</style>*@
