+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/covid/PlaceOrder.cshtml
@@ -1,4 +1,4 @@
-﻿@model Momentarily.ViewModels.Models.CovidGoodViewModel
+﻿<!--@model Momentarily.ViewModels.Models.CovidGoodViewModel
 @{
     Layout = "~/Areas/Frontend/Views/Shared/_LayoutNew.cshtml";
 }
@@ -396,9 +396,9 @@
             </div>
         </div>
     </div>
-</footer>
+</footer>-->
 <!---->
-<div class="copyright">
+<!--<div class="copyright">
     <div class="container">
         <div class="row">
             <div class="col-sm-6">
@@ -648,4 +648,4 @@
     });
 
 
-</script>
+</script>-->
