+++ Momentarily/Momentarily.Web/Areas/Frontend/Views/covid/_OrderPaymentResult.cshtml
@@ -1,4 +1,4 @@
-﻿@model Momentarily.ViewModels.Models.OrderPayResult
+﻿@*@model Momentarily.ViewModels.Models.OrderPayResult
     @if (!string.IsNullOrWhiteSpace(Model.Message))
     {
         <div class="alerts">
@@ -11,5 +11,5 @@
                 <div class="alert-error successMessage text-center">@Model.Message</div>
             }
         </div>
-    }
+    }*@
