+++ Momentarily/Momentarily.Web/Web.config
@@ -45,8 +45,9 @@
   </log4net>
   <connectionStrings>
     <!--<add name="Mssqlconnectionstring" connectionString="Data Source=eva.empeek.net,55483;Initial Catalog=Momentarily;User ID=;Password=;MultipleActiveResultSets=True;" providerName="System.Data.SqlClient" />-->
-    <add name="Mssqlconnectionstring" connectionString="Data Source=DESKTOP-R5VAJ5R;Initial Catalog=Momentarily;integrated security=true;MultipleActiveResultSets=True;" providerName="System.Data.SqlClient" />
-    <!--<add name="Mssqlconnectionstring"
+    <!--<add name="Mssqlconnectionstring" connectionString="Data Source=DESKTOP-R5VAJ5R;Initial Catalog=Momentarily;integrated security=true;MultipleActiveResultSets=True;" providerName="System.Data.SqlClient" />-->
+	  <add name="Mssqlconnectionstring" connectionString="Data Source=localhost;Initial Catalog=Momentarily;integrated security=true;MultipleActiveResultSets=True;" providerName="System.Data.SqlClient" />
+	  <!--<add name="Mssqlconnectionstring"
                connectionString="data source=.;
    Initial Catalog=Momentarily;Integrated Security=True;"
                providerName="System.Data.SqlClient" />-->
